+++ Results/TieResult.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+require_once 'Signs/ResultEngine.php';
+
+class TieResult extends ResultEngine
+{
+    public static function getMessage(): string
+    {
+        return "It's a tie!" . PHP_EOL;
+    }
+}
