--- SignInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-interface SignInterface
-{
-    public function beats(SignInterface $sign): Result;
-}
