+++ Signs/Rock.php
@@ -3,17 +3,17 @@
 declare(strict_types=1);
 
 require_once 'SignInterface.php';
-require_once 'TieResult.php';
-require_once 'WinResult.php';
-require_once 'LoseResult.php';
+require_once 'Results/TieResult.php';
+require_once 'Results/WinResult.php';
+require_once 'Results/LoseResult.php';
 
-class Paper extends Result implements SignInterface
+class Rock extends ResultEngine implements SignInterface
 {
     protected array $beateable = [
-        Rock::class
+        Scissors::class
     ];
 
-    public function beats(SignInterface $element): Result
+    public function beats(SignInterface $element): ResultEngine
     {
         if ($this instanceof $element) {
             return new TieResult();
@@ -22,6 +22,7 @@ class Paper extends Result implements SignInterface
         if (in_array(get_class($element), $this->beateable)) {
             return new WinResult();
         }
+
         return new LoseResult();
     }
 }
