+++ Signs/SignInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+interface SignInterface
+{
+    public function beats(SignInterface $sign): ResultEngine;
+}
