+++ Signs/SignsCollection.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+class SignsCollection
+{
+    private array $signs = [];
+
+    public function addSigns()
+    {
+        
+    }
+}
