+++ index.php
@@ -2,12 +2,17 @@
 
 declare(strict_types=1);
 
-require_once 'Rock.php';
-require_once 'Paper.php';
-require_once 'Scissors.php';
+require_once 'Signs/Rock.php';
+require_once 'Signs/Paper.php';
+require_once 'Signs/Scissors.php';
+require_once 'Signs/SignsCollection.php';
 
 $rock = new Rock();
 $paper = new Paper();
 $siccors = new Scissors();
 
-var_dump($rock->beats($paper));
+$outcome = $rock->beats($siccors);
+
+$result = new $outcome;
+
+echo $result->getMessage();
